# backend/docs/swagger.yaml
basePath: /
definitions:
...
  job.CreateJobRequest:
    properties:
      body:
        type: string
      description:
        type: string
      endpoint:
        type: string
      headers:
        additionalProperties: {}
        type: object
      method:
        enum:
        - GET
        - POST
        - PUT
        - DELETE
        - PATCH
        type: string
      name:
        type: string
      schedule:
        type: string
    required:
    - endpoint
    - headers
    - method
    - name
    - schedule
    type: object
  job.CreateJobResponse:
    properties:
      body:
        type: string
      description:
        type: string
      endpoint:
        type: string
      headers:
        additionalProperties: {}
        type: object
      id:
        type: string
      method:
        type: string
      name:
        type: string
      schedule:
        type: string
    type: object
  job.ErrorResponse:
    properties:
      error:
        type: string
    type: object
  job.JobResponse:
    properties:
      body:
        type: string
      description:
        type: string
      endpoint:
        type: string
      headers:
        additionalProperties: {}
        type: object
      id:
        type: string
      method:
        type: string
      name:
        type: string
      schedule:
        type: string
    type: object
...
host: localhost:5000
info:
  contact: {}
  description: This is a simple API for scheduling and executing jobs.
  title: Kron API
  version: "1.0"
paths:
...
  /api/job/all:
    get:
      description: Find all created jobs
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/job.JobResponse'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/job.ErrorResponse'
      summary: Find all jobs
  /api/job/create:
    post:
      consumes:
      - application/json
      description: Creates a new job.
      parameters:
      - description: Job
        in: body
        name: job
        required: true
        schema:
          $ref: '#/definitions/job.CreateJobRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Success
          schema:
            $ref: '#/definitions/job.CreateJobResponse'
        "400":
          description: Invalid request
          schema:
            $ref: '#/definitions/job.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/job.ErrorResponse'
      summary: Create a job
...
swagger: "2.0"